+++ pennylane_lightning/core/bindings/BindingsMPI_nb.hpp
@@ -39,6 +39,7 @@
 #include "AdjointJacobianGPUMPI.hpp"
 #include "JacobianDataMPI.hpp"
 #include "LGPUBindingsMPI.hpp"
+#include "MPIManagerGPU.hpp"
 #include "MeasurementsGPUMPI.hpp"
 #include "ObservablesGPUMPI.hpp"
 
@@ -48,14 +49,16 @@ using namespace Pennylane::LightningGPU;
 using namespace Pennylane::LightningGPU::Algorithms;
 using namespace Pennylane::LightningGPU::Observables;
 using namespace Pennylane::LightningGPU::Measures;
+using namespace Pennylane::LightningGPU::Util;
+
 } // namespace
 /// @endcond
 
 #elif _ENABLE_PLKOKKOS == 1
 
 #include "AdjointJacobianKokkosMPI.hpp"
-#include "JacobianDataMPI.hpp"
-#include "LKokkosBindingsMPI.hpp"
+#include "LKokkosBindingsMPI_nb.hpp"
+#include "MPIManagerKokkos.hpp"
 #include "MeasurementsKokkosMPI.hpp"
 #include "ObservablesKokkosMPI.hpp"
 
@@ -65,6 +68,8 @@ using namespace Pennylane::LightningKokkos;
 using namespace Pennylane::LightningKokkos::Algorithms;
 using namespace Pennylane::LightningKokkos::Observables;
 using namespace Pennylane::LightningKokkos::Measures;
+using namespace Pennylane::LightningKokkos::Util;
+
 } // namespace
 /// @endcond
 
@@ -89,13 +94,11 @@ template <class StateVectorT> void registerObservablesMPI(nb::module_ &m) {
         typename StateVectorT::PrecisionT; // Statevector's precision.
     using ComplexT =
         typename StateVectorT::ComplexT; // Statevector's complex type.
-    using ParamT = PrecisionT;           // Parameter's data precision
 
     const std::string bitsize =
         std::is_same_v<PrecisionT, float> ? "64" : "128";
 
-    using arr_c = nb::ndarray<std::complex<ParamT>, nb::c_contig>;
-    using arr_r = nb::ndarray<ParamT, nb::c_contig>;
+    using ArrCT = nb::ndarray<std::complex<PrecisionT>, nb::c_contig>;
     using ObservableT = Observable<StateVectorT>;
     using ObsPtr = std::shared_ptr<ObservableT>;
 
@@ -170,7 +173,7 @@ template <class StateVectorT> void registerObservablesMPI(nb::module_ &m) {
     // Register HamiltonianMPI class
     class_name = "HamiltonianMPIC" + bitsize;
     nb::class_<HamiltonianMPI<StateVectorT>, ObservableT>(m, class_name.c_str())
-        .def(nb::init<const std::vector<ParamT> &,
+        .def(nb::init<const std::vector<PrecisionT> &,
                       const std::vector<ObsPtr> &>())
         .def("__repr__", &HamiltonianMPI<StateVectorT>::getObsName)
         .def("get_wires", &HamiltonianMPI<StateVectorT>::getWires,
@@ -189,8 +192,8 @@ template <class StateVectorT> void registerObservablesMPI(nb::module_ &m) {
 
 #if _ENABLE_PLGPU == 1
     using SparseIndexT =
-        typename std::conditional<std::is_same<ParamT, float>::value, int32_t,
-                                  int64_t>::type;
+        typename std::conditional<std::is_same<PrecisionT, float>::value,
+                                  int32_t, int64_t>::type;
 
     // Register SparseHamiltonianMPI class
     class_name = "SparseHamiltonianMPIC" + bitsize;
@@ -229,7 +232,6 @@ template <class StateVectorT> void registerObservablesMPI(nb::module_ &m) {
 template <class StateVectorT, class PyClass>
 void registerBackendAgnosticMeasurementsMPI(PyClass &pyclass) {
     using PrecisionT = typename StateVectorT::PrecisionT;
-    using ParamT = PrecisionT;
     using ObsPtr = std::shared_ptr<Observable<StateVectorT>>;
 
     pyclass
@@ -259,7 +261,9 @@ void registerBackendAgnosticMeasurementsMPI(PyClass &pyclass) {
                 std::size_t num_shots) {
                  return createNumpyArrayFromVector<std::size_t>(
                      M.generate_samples(num_shots), num_shots, num_wires);
-             });
+             })
+        .def("set_random_seed", [](MeasurementsMPI<StateVectorT> &M,
+                                   std::size_t seed) { M.setSeed(seed); });
 }
 
 /**
@@ -304,9 +308,8 @@ void registerBackendAgnosticAlgorithmsMPI(nb::module_ &m) {
         typename StateVectorT::PrecisionT; // Statevector's precision
     using ComplexT =
         typename StateVectorT::ComplexT; // Statevector's complex type
-    using ParamT = PrecisionT;           // Parameter's data precision
 
-    using arr_c = nb::ndarray<std::complex<ParamT>, nb::c_contig>;
+    using ArrCT = nb::ndarray<std::complex<PrecisionT>, nb::c_contig>;
 
     const std::string bitsize =
         std::is_same_v<PrecisionT, float> ? "64" : "128";
@@ -320,7 +323,7 @@ void registerBackendAgnosticAlgorithmsMPI(nb::module_ &m) {
     class_name = "OpsStructMPIC" + bitsize;
     nb::class_<OpsData<StateVectorT>>(m, class_name.c_str())
         .def(nb::init<const std::vector<std::string> &,
-                      const std::vector<std::vector<ParamT>> &,
+                      const std::vector<std::vector<PrecisionT>> &,
                       const std::vector<std::vector<std::size_t>> &,
                       const std::vector<bool> &,
                       const std::vector<std::vector<ComplexT>> &>())
@@ -349,12 +352,12 @@ void registerBackendAgnosticAlgorithmsMPI(nb::module_ &m) {
            const std::vector<std::vector<PrecisionT>> &ops_params,
            const std::vector<std::vector<std::size_t>> &ops_wires,
            const std::vector<bool> &ops_inverses,
-           const std::vector<arr_c> &ops_matrices,
+           const std::vector<ArrCT> &ops_matrices,
            const std::vector<std::vector<std::size_t>> &ops_controlled_wires,
            const std::vector<std::vector<bool>> &ops_controlled_values) {
             std::vector<std::vector<ComplexT>> conv_matrices =
                 Pennylane::NanoBindings::Utils::convertMatrices<ComplexT,
-                                                                ParamT>(
+                                                                PrecisionT>(
                     ops_matrices);
             return OpsData<StateVectorT>{ops_name,
                                          ops_params,
@@ -376,7 +379,7 @@ void registerBackendAgnosticAlgorithmsMPI(nb::module_ &m) {
              "Adjoint Jacobian method.")
         .def(
             "batched",
-            [](AdjointJacobianMPI<StateVectorT> &self, StateVectorT &sv,
+            [](AdjointJacobianMPI<StateVectorT> &self, const StateVectorT &sv,
                const std::vector<std::shared_ptr<Observable<StateVectorT>>>
                    &observables,
                const OpsData<StateVectorT> &operations,
@@ -389,7 +392,7 @@ void registerBackendAgnosticAlgorithmsMPI(nb::module_ &m) {
                 const JacobianDataMPI<StateVectorT> jd{
                     operations.getTotalNumParams(), sv, observables, operations,
                     trainableParams};
-                adjoint_jacobian.adjointJacobian_serial(std::span{jac}, jd);
+                self.adjointJacobian_serial(std::span{jac}, jd);
 #elif _ENABLE_PLKOKKOS == 1
                 const JacobianData<StateVectorT> jd{
                     operations.getTotalNumParams(),
@@ -398,11 +401,8 @@ void registerBackendAgnosticAlgorithmsMPI(nb::module_ &m) {
                     observables,
                     operations,
                     trainableParams};
-                adjoint_jacobian.adjointJacobian(std::span{jac}, jd, sv);
-
+                self.adjointJacobian(std::span{jac}, jd, sv);
 #endif
-                self.adjointJacobian_serial(std::span{jac}, jd);
-
                 return createNumpyArrayFromVector<PrecisionT>(std::move(jac));
             },
             "Batch Adjoint Jacobian method.");
@@ -414,29 +414,8 @@ void registerBackendAgnosticAlgorithmsMPI(nb::module_ &m) {
  * @param m Nanobind module
  */
 inline void registerInfoMPI(nb::module_ &m) {
-    nb::class_<MPIManager>(m, "MPIManager")
-        .def(nb::init<>())
-        .def(nb::init<MPIManager &>())
-        .def("Barrier", &MPIManager::Barrier)
-        .def("getRank", &MPIManager::getRank)
-        .def("getSize", &MPIManager::getSize)
-        .def("getSizeNode", &MPIManager::getSizeNode)
-        .def("getTime", &MPIManager::getTime)
-        .def("getVendor", &MPIManager::getVendor)
-        .def("getVersion", &MPIManager::getVersion)
-        // Template version with explicit type constraints
-        .def(
-            "Scatter",
-            []<typename T>(MPIManager &mpi_manager,
-                           nb::ndarray<std::complex<T>, nb::c_contig> &sendBuf,
-                           nb::ndarray<std::complex<T>, nb::c_contig> &recvBuf,
-                           int root) {
-                auto send_ptr = sendBuf.data();
-                auto recv_ptr = recvBuf.data();
-                mpi_manager.template Scatter<std::complex<T>>(
-                    send_ptr, recv_ptr, recvBuf.size(), root);
-            },
-            "MPI Scatter for complex arrays.");
+    // This function is now empty - MPI manager registration moved to
+    // backend-specific
 }
 
 /**
@@ -479,7 +458,7 @@ template <class StateVectorT> void lightningClassBindingsMPI(nb::module_ &m) {
     auto pyclass_measurements =
         nb::class_<MeasurementsMPI<StateVectorT>>(m, class_name.c_str());
 
-    pyclass_measurements.def(nb::init<const StateVectorT &>());
+    pyclass_measurements.def(nb::init<StateVectorT &>());
     registerBackendAgnosticMeasurementsMPI<StateVectorT>(pyclass_measurements);
     registerBackendSpecificMeasurementsMPI<StateVectorT>(pyclass_measurements);
 
@@ -499,8 +478,6 @@ void registerLightningClassBindingsMPI(nb::module_ &m) {
         using StateVectorT = typename TypeList::Type;
         lightningClassBindingsMPI<StateVectorT>(m);
         registerLightningClassBindingsMPI<typename TypeList::Next>(m);
-        nb::exception<Pennylane::Util::LightningException>(
-            m, "LightningExceptionMPI");
     }
 }